split files into libs
